+++ src/main/java/net/fhirfactory/pegacorn/hestia/audit/im/processingplant/configuration/HestiaAuditIMTopologyFactory.java
@@ -1,10 +1,10 @@
-package net.fhirfactory.pegacorn.hestia.im.processingplant.configuration;
+package net.fhirfactory.pegacorn.hestia.audit.im.processingplant.configuration;
 
 import net.fhirfactory.pegacorn.deployment.properties.configurationfilebased.common.segments.ports.interact.StandardInteractClientPortSegment;
 import net.fhirfactory.pegacorn.deployment.topology.factories.archetypes.fhirpersistence.im.FHIRIMSubsystemTopologyFactory;
 import net.fhirfactory.pegacorn.deployment.topology.model.nodes.*;
 import net.fhirfactory.pegacorn.deployment.topology.model.nodes.common.EndpointProviderInterface;
-import net.fhirfactory.pegacorn.hestia.im.common.HestiaIMNames;
+import net.fhirfactory.pegacorn.hestia.audit.im.common.HestiaIMNames;
 import net.fhirfactory.pegacorn.util.PegacornEnvironmentProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
